+++ src/main/java/reactor/core/publisher/FluxOnBackpressureBuffer.java
@@ -63,7 +63,7 @@ final class FluxOnBackpressureBuffer<O> extends FluxSource<O, O> implements Fuse
 
 	@Override
 	public long getPrefetch() {
-		return Long.MAX_VALUE;
+		return Integer.MAX_VALUE;
 	}
 
 	static final class BackpressureBufferSubscriber<T>
@@ -360,7 +360,7 @@ final class FluxOnBackpressureBuffer<O> extends FluxSource<O, O> implements Fuse
 
 		@Override
 		public long getCapacity() {
-			return Long.MAX_VALUE;
+			return Integer.MAX_VALUE;
 		}
 
 		@Override
