+++ src/main/java/reactor/core/publisher/LambdaSubscriber.java
@@ -148,7 +148,7 @@ final class LambdaSubscriber<T>
 
 	@Override
 	public long getCapacity() {
-		return Long.MAX_VALUE;
+		return Integer.MAX_VALUE;
 	}
 
 	@Override
